standalone-sakhi-ai/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
google-generativeai==0.3.2
chromadb==0.4.22
pypdf==3.17.4
pdfplumber==0.10.3
python-dotenv==1.0.0
requests==2.31.0
speechrecognition==3.10.1
pyttsx3==2.90
pyaudio==0.2.14
soundfile==0.12.1
pydub==0.25.1
numpy==1.24.3
scikit-learn==1.3.2
textstat==0.7.3
nltk==3.8.1
transformers==4.36.2
torch==2.1.2
pillow==10.1.0
opencv-python==4.8.1.78
face-recognition==1.3.0
streamlit==1.29.0
plotly==5.17.0
pandas==2.1.4
openai==1.6.1
elevenlabs==0.2.26
whisper==1.1.10
watchdog==3.0.0
aiofiles==23.2.1
asyncio==3.4.3
websockets==12.0
sqlite3
uuid
datetime
json
os
logging
pathlib
typing
dataclasses
threading
queue
time
hashlib
base64
io
re
math